REDVM-484: change blobstore location
we lost access to our service backup blobstore
and need to move everything to a blobstore that
we own.

this updates our config.yml to point to the new
blobstore.

also, none of our final_builds or releases are
resolvable because of the blobstore changes,
therefore those references must be deleted.
this is why final builds and releaese have
all been deleted.

after these chnages are made, we are able to
upload new releases into our new blobstore